dumb_photo_idea_illustraton/requirements.txt
bchlib==2.1.3
certifi==2025.7.14
charset-normalizer==3.4.2
filelock==3.18.0
fsspec==2025.7.0
ftfy==6.3.1
hf-xet==1.1.5
huggingface-hub==0.34.3
idna==3.10
ImageHash==4.3.2
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
open_clip_torch==3.0.0
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
PyWavelets==1.6.0
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
safetensors==0.5.3
scipy==1.13.1
sympy==1.14.0
timm==1.0.19
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
wcwidth==0.2.13
